Move getUserNameWithoutDomain to modules/core
